Fix random failure spec
This is about the random failure specs for interruption with Ctrl-C
https://travis-ci.org/bbatsov/rubocop/jobs/6877252

The specs were testing real behavior by running rubocop
as a child process. However it seems to be unstable
and also cannot be measured in coverage report.
So rewrote them to run in thread with some stubs and mocks.