Force JRuby not to select working directory as temporary directory
JRuby selects a temporary directory from the candidate directories,
which includes the ordinary /tmp and also the current working directory,
by checking if they are world writable or not. If /tmp is world writable
and the current working directory is not, the latter is chosen as a
temporary directory. This affects to behaviors of Tempfile and
Dir.mktmpdir. This is the reason that the spec context
"isolated enviroment" was actually not isolated on JRuby.

This is related to the following issue:
https://github.com/jruby/jruby/issues/405