Fix Parser dependency to 2.0.0.beta6
Parser 2.0.0 is still beta release.
This means backward compatibility is not always kept,
and RuboCop may break if the dependency version is unfixed.
In development of RuboCop, we need to differentiate
the cause of broken specs. Otherwise we cannot rely on specs.

When a new beta version of Parser is released,
we can check whether RuboCop works properly with the new version
and fix if needed, then update the version of the dependency.

After Parser 2.0.0 official release is done,
we would specify the dependency as range like '~> 2.0'.