Remove the BlockAlignSchema config option from EndAlignment
Having two different ways to align the block's end complicates a lot
the code and the specs. Basically most editors align the block's end
with the start of the expression where the block is defined.
This is how we handle the alignment now and this was previously our
default option as well.