Disable the Coveralls reports
Seems that there is an error when generating the Coveralls report
and this results in a check in the colorize gem to throw IOError.
This error results in a exit code of 1 and marks the build as failed
even that we have all specs passing.

I believe there is an error with the Coveralls API, but I couldn't
reproduce the error message locally, so I'm disabling the service
in attempt to restore our builds on Travis.