Fix bug concerning table alignment of colon separated hash entries.
Table alignment required the colon to be detached from the key, which
is a syntax error.

Also, colon and hash rocket are not mixed in the specs anymore. We don't
have full support for that kind of mix, so let's not make readers believe
that we do.

I have some changes lined up to take care of concerns raised in #471,
but first I want to add this bugfix.