Make AccessControl indent depth configurable
Add the ability for the `AccessControl` cop to respect the new
`IndentDepth` configuration parameter, which specifies whether
`private`/`protected` keywords are indented to the same indentation as
the class or methods within that class.

Valid values for `IndentDepth` are 'class' and 'method'. The default
is 'method', which maintains existing behavior.

The justification for allowing this to be configurable is that some
codebases prefer indenting these modifiers at the same level as the
containing class. The reason for this is that it visually separates
"sections" of methods with different visibilities, and is easier pick
out visually when viewing a file as there is a horizontal shift in
position from that of the rest of the method `def`initions, making it
stand out.