Expand ~ in inherit_from paths to home directory
A common feature in the configuration of many tools is the ability to
use the `~` character to refer to a user's home directory.

Since arbitrary interpolation in Rubocop config files is not yet
supported, implementing this seems like a simple step to providing this
kind of flexibility, which is especially useful for development
environments which have a general set of configuration settings in the
home directory, along with repo-specific configurations in each repo.