Let IndentationWith accept alignment with variable in var = if.
This makes the cop accept indentation of an if body relative to the
left hand side of an assignment. No configuration parameter is added,
we just allow two styles. The reason is that EndAlignment now has a
configuration parameter for choosing between the two styles, and it's
enough to make that choice in one place.