Fix problem with [Corrected] tag sometimes missing
At this point, autocorrect is a "do or do not do, there is no try"
kind of function. So in the special cases where a correction is
avoided because of interference from some other correction, it's
really only postponed and will be done sooner or later. The
`corrected` attribute of `Offence` needs to reflect that and be
set to `true` if the cop supports autocorrect. Otherwise, we
don't get the `[Corrected]` tag in the report.