Whitelist the Rakefile and Gemfile from name check
When building a Ruby project it's very common that you'll have a Rakefile and a Gemfile. These files are normally capitalized, which violates the default Rubocop filename checks. This changeset adds a whitelist for base filenames that should *NOT* be included in the file_name checks.

This works by adding a constant to the FileName class, WHITELIST, which is an Array that contains the whitelisted basenames. Later on, it will only add an offense if the basename is not snake case, and the basename is not Rakefile or Gemfile.

I've updated the tests for this specific use case and have confirmed that all tests pass on my local system.

-Tim