[Fix #1111] Fix reading of non-UTF-8 files in EndOfLine
The problem was that we used buffer.source.encoding as the file's
encoding when reading it again, but Parser has re-encoded the buffer
so it's always UTF-8. By first reading the file as ascii-8bit
(binary) encoding, and then re-encoding it based on Parser's logic
for recognizing encodings, we should be able to match /\r$/ always.