Store encoding error in ProcessedSource and merge SourceParser into it
By storing the encoding error on parsing in ProcessedSource rather than
raising it, the offense emitting logic in Runner can be moved to
Cop::Team.

https://github.com/bbatsov/rubocop/pull/900#discussion-diff-10940125

Without merging the two classes, to store the error in ProcessedSource,
we would need to pass 6 arguments to ProcessedSource.new.
Also, since some of the attributes are dependent on others, it seems
that how the attributes are generated/dependent should be a knowledge of
ProcessedSource itself.