Restrict valid schemes for LineLength/AllowURI
Per #1335, #1297, the existing way of matching URIs was too permissive
and allowed a lot of lines with URI-like content (e.g. Ruby namespaces)
to be ignored by the LineLength cop.

To work around this, a whitelist of schemes will stop the vast majority
of false positives.

URISchemes was added as a configuration option for flexibility.

Closes #1335

Fixed some missing test coverage that resulted from the above changes.
The configurable URISchemes caused an existing test to miss the line of
code it should have covered.