Regex is slow, start_with? is fast
The `[qQ]?` part of the regex was superfluous, as was the `i` flag.

Some code -- I am not sure which -- has been changed, so that the
`<<-` part of the regex is no longer necessary. There specs that cover
the heredocs are still in place though (and the code works with the
`<<` style heredoc as well, by the way).

What was left was `src =~ /^(%|\?)/` which is out-performed by
`String#start_with?`