[Fix #1368] MultilineOperationIndentation's handling of RSpec code
Change what is regarded as the left hand side, or base
indentation, that the right hand side is compared against.

Typical RSpec code with expect ... to change ... from ... to
is a good complicated example. It's not self evident that
the indentation should be just this way. I've made a choice.