Do not use the syck YAML engine
We started using the syck engine in some cases to
solve a roblem that
a) we can't reproduce now
b) we can solve in another way

This is the other solution. To have special logic
for empty YAML files.

The reason why we want to avoid syck is that it
gets into encoding problems with the configuration
of Style/WordArray, which contains a regular
expression.