Show decimals for Metrics/AbcSize:Max if given in config
The reporting of the offense should include a decimal
if set like that in a configuration file.

And increase the number of significant digits reported
for the size. Otherwise we get output such as [44/44]
when the size is just above 44.

Allowing a little bit of duplication makes the specs
more readable than with the method registers_offense.