[Fix #1425] Find included files in a simpler way
Only look in the "base" configuration when finding the
target files, not in the configuration associated with
each file under consideration. This removes a feature
that has existed for a long time, but hopefully isn't
used by many people (the ability to put AllCops/
Include parameters in .rubocop.yml files in
subdirectories). We didn't cover the existing behavior
in specs, but the new behavior is covered.