[Fix #1500] Remove double handling of elsif in IndentationWidth
The fact that elsif bodies were checked twice could cause
auto-correction to become unstable and indent indefinitely.
The reason was that one check was relative to the if keyword,
the other to the elsif keyword. So if these were not aligned,
and the ElseAlignment cop was not enabled, there would always
be an offense to correct.