Make coverage measurements work again
The order of require calls in spec_helper.rb is important. It had
been changed in order to make it possible to have test cops under
the support directory. This made coverage plummet to around 80%.
This change restores the old order and finds a way to test infinite
loops without special test cops.