change strategy for removing begin/end
This commit updates the strategy for removing the extra begin/end.
In the previous version, when a heredoc ended the child node,
the end of the range to be replaced was returned as the first line
of the heredoc, not the last line of the heredoc. I believe this
may be an issue with the parse library.