[Fix #1865] Identify unnecessary rubocop:disable comments
Keep all offenses around, including locally disabled ones,
until reporting time.

Let the formatters do the filtering by adding a new method
wanted_offenses in the public API of BaseFormatter. The
default behavior is to not keep the comment-disabled
offenses, which makes the change backwards compatible.

Likewise, the public API of Offense is amended with a
disabled? method, but the corrected/corrected? methods
stay compatible.

DisabledLinesFormatter outputs a red "(unnecessary)"
after lines with a disabled cop and a line range that
didn't have any offenses for that cop.