Account for multi-line hashes in method calls.
The `TrailingComma` cop treats each method parameter as an individual
item when looking for multi-line method calls. If there is only a single
item and that item is a hash the cop does not check if that hash spans
multiple lines.

This fix ensures that method calls with a single hash parameter can still
qualify as spanning multiple lines if the hash itself spans multiple
lines.