[Fix #1918] Add config param AllCops:DisabledByDefault
By disabling cops in default configuration we make the
selection process "opt-in". In order for this feature
to work well together with --auto-gen-config (and any
pre-existing config file), we allow the mere existence
of configuration for a cop in a user configuration file
to make the cop enabled (the Enabled parameter does not
have to be set).