[Fix #2045] Fix error on class method definition with modifier
We were assuming that method body is always the third element of a
method definition node

```
_method_name, _args, body = *args.first
```

and here is an example how args.first looks in case of regular methods:
```
(def :test
  (args)
  (send nil :body))
```

but in case of class methods, for example in this code
```
private_class_method def self.bar
```

there is an extra `(self)` element:
```
(defs
  (self) :bar
  (args)
  (send nil :body))
```

So the body becomes the fourth element instead.

Always taking the last element takes care
of the issue.

I've added some tests for cases where there was a def test without a defs counterpart.