[Fix #2023] Avoid double auto-correct in IndentationWidth
IndentationWidth can auto-correct a node spanning
multiple lines. This is a useful feature, because
it takes care of `end` keywords, changing their
indentation together with the chunk that they
surround. The `end` keywords are not auto-corrected
by default in the cops that specifically deal with
them.

The problem with this feature in IndentationWidth
is that incorrect indentation can be found within
a node that is being corrected. This results in
corrupted code and must be avoided. The next round
in the iterative auto-correction process will
pick up the skipped correction.