Fix several issues in Lint/FormatParameterMismatch
* Fixes #2123 where dynamic widths were being ignored.

The regex has been updated to detect dymanic widths. A dynamic
width duplicates the number of parameters needed for each expected
field:

  format("%*d", max_width, id)

* Fixes #2116 named params

We ignore the case when a hash is passed as the source for the
fields. It is not necessary as Ruby already checks that and will
launch an exception if it doesn't match.

* Missng 'A' formatter

The regex has been updated to support the A formatter, which was
missing.

* HEREDOC support

The presence of heredocs with format has been fixed. Previously,
it threw off the number of parameters found.