[Fix #2109] Consider indentation in ExtraSpacing
When searching for possibly aligned tokens, if none is
found using the normal search algorithm that looks on the
nearest lines, skipping empty and comment lines, do
another search where only lines with the same indentation
are considered.

This must be an addition to the existing algorithm. The
old stuff is still needed.