Substitute slow regex for something faster
Somehow, the `\s*` part of the regex makes it really slow, I guess
it's looking through the entire "source_before_end" string and not
just looking backwards from the end of the string.

This approach will partition the string and check if the last line is
empty.

Informal benchmarking shows around 30% speed increase on
`Lint/BlockAlignment` and `Style/IndentationWidth`, while
`Style/MultilineOperationIndentation` seems to be around 45% faster.
The exact numbers obviously depend a whole lot on the input.