Make relative Excludes work in .rubocop_todo.yml
When Exclude properties are read, we make the paths absolute
so that matching against them will be easy regardless of
current directory. The interpretation of relative paths is
different when they are specified in .rubocop.yml compared
to, for example, default.yml. What we forgot is that the
same goes for .rubocop_todo.yml too.