Always allow a StyleGuide parameter
StyleGuide parameters are currently only allowed for cops which already
have one in the default configuration. There is a validation warning when
a StyleGuide parameter is found in the local configuration.

It can be useful to add a link to a specific style guide providing some
rationale for the configuration, so this patch allows to always add a
StyleGuide parameter.