Don't warn for using eval() on a literal string
Using eval on a string literal, with no computed values interpolated into it, is not a
"security risk" and there is no need to warn about it.

Why would anyone do this, anyways? Here is an example:

    if RUBY_VERSION.to_f >= 1.9
      eval "code which uses 1.9 syntax..."
    end

This allows an older interpreter to still load and parse the source file.