Lint/FormatParameterMismatch ignores % if LHS or RHS is not a literal
If the LHS is not a literal, it could evaluate to a format string with any
number of arguments, so we can't count an offense for that. On the other hand,
if the RHS is not a literal, it could evaluate to an array with any number
of arguments, so we can't count an offense for that either.

Fixes #2307.