[Fix #2003] Let --auto-correct disable Lint/UnneededDisable
The problem is this. When Runner calls FormatterSet#file_started,
it gives the cop_disabled_line_ranges and it's stored in
FormatterSet. Then Runner starts its inspection loop for the file
and each time the file is updated (by autocorrect), a new
processed_source object is created with new
cop_disabled_line_ranges. But the reference in FormatterSet is
still to the original line ranges.

One possible solution could be to try to update the
processed_source object instead. That seems hard. This solution
takes the easy way out, just disabling the UnneededDisable cop in
this tricky situation.