[Fix #2003] Handle --auto-correct in Lint/UnneededDisable
The problem is this. When Runner calls FormatterSet#file_started,
it gives the cop_disabled_line_ranges and comments and they are
stored in FormatterSet. Then Runner starts its inspection loop
for the file and each time the file is updated (by autocorrect),
a new processed_source object is created with new
cop_disabled_line_ranges and comments. But the references in
FormatterSet are still to the original ranges and comments.

The solution is to send delegator objects to file_started. These
delegators are updated each time the code is parsed again after
correction, so that the disabled lines and comments that
Lint/UnneededDisable uses are the latest, rather than the ones
given before inspection started.