Call :expression on node instead of :keyword loc
This changes `RuboCop::Cop::CodeLength`s `check_code_length` method to
return `node.loc` instead of `:keyword`.

When `:keyword` is passed in it calls `node.loc.keyword` which returns
the keyword, eg: `class`, `def`, or `module` and only reports the line
with that keyword as the beginning and end lines.

This passes `:expression` instead which will return the entire block
instead of the defining line.