Make inherit_from override inherit_gem
The problem was discovered in #2411, but this change does not
fix the problem reported there.

Make configuration loading work as described in the README regarding
precedence of `.rubocop.yml`, `inherit_from`, and `inherit_gem`.

Remove `rubocop: config/default.yml` from the example in the
README as it makes no sense. All configurations are based on the
`default.yml` of the `rubocop` being run.