Fix ProgressFormatter not respecting `--no-color`
The coloured dots appear in the output even when the CLI switch to not use colour has been passed. At the point when the formatter is initialised `Rainbow.enabled` has been disabled globally if we want no colour so we can check this too before opting for the `GREEN_DOT`.