New Cop Style/MultilineArrayBraceLayout
This cop is inspired by a comment in the `Style/LineBreaks` cop PR:

https://github.com/bbatsov/rubocop/pull/2277#issuecomment-148615911

This cop checks that the closing brace in an array literal is symmetrical
with respect to the opening brace and the array elements. I am very open
to suggestions on naming and ways to clarify the offense messages.

    # bad
    [ :a,
      :b
    ]

    # bad
    [
      :a,
      :b ]

    # good
    [ :a,
      :b ]

    #good
    [
      :a,
      :b
    ]

My intention is to expand this to account for hashes, method parameters
and method arguments once the language and approach is validated.