[Fix #2452] Style/TrailingComma and single-line hashes in method calls
`Style/TrailingComma` always assumes that all the elements of a hash
count towards the arguments of a method call for the purposes of
determining if the method-call is multi-line.

This is a problem because each pair of consecutive arguments is checked
to determine if any arguments fall on the same line. Since all the
elements of a single-line hash always fall on the same line, this check
will always fail.

Visually it makes sense to consider a single-line hash as a single
argument in a method call.

Example:

```
method(
  { a: 1, b: 2 },
  { a: 2, b: 3 },
)
```

I have corrected this behavior by only counting the elements of a hash
if the hash itself is a multi-line hash.