[Fix #2066] Style/TrivialAccessors doesn't flag method defs inside instance_eval
In some cases, `instance_eval` may be called on a class object. In that case,
`attr_reader`, etc. can be used in the block. But we can't detect that reliably.
Therefore, don't flag what appear to be trivial accessor defs if they are
inside a call to `instance_eval`. However, if there is a `class` or `class <<`
block nested inside the call to `instance_eval`, flag defs inside it.

Detection of whether a def is inside a module or not is also slightly more
robust now.