ConfigurableEnforcedStyle doesn't write arrays into auto-genned config
ConfigurableEnforcedStyle recently received an upgrade which allows it to work
with 3 or more supported styles, and intelligently select the style which
matches EVERYTHING found in the code (if there is such a style).

When there is more than one style which could potentially be a match, it does
this by maintaining an array of all the candidates which match everything seen
so far.

However, when the end of all the files to be scanned is reached, there may still
be 2 or more candidates remaining. If so, make sure that only one of them is
written into .rubocop.yml. (The choice is arbitrary.)