If config file validation fails, fail hard with an error message
This will force users to pay attention to validation errors, rather letting
them quietly pass by in a CI console which no-one may actually be looking at.

It also means that for validation warnings, as opposed to errors, more than one
warning can be displayed if there is more than one problem in a configuration
file. Previously, it was only possible for one validation warning to be shown.