Cache offenses only, not comments and disabled line ranges
The cached disabled line ranges and comments are used to add UnneededDisable
offenses to the cached offenses. But this process is deterministic, so we can
simply do it *before* caching and then just cache the resulting offenses.

This greatly simplifies the caching code. As a bonus, the caching should be
more effective, since more work is done once only and then reused, rather
than being repeated every time the cached data is retrieved.