[Fix #2574] Fix 'variable' style for Lint/EndAlignment
We had no tests checking whether `Lint/EndAlignment`'s `variable` style worked
or not when a node with `end` is not on the RHS of an assignment. So it was
never noticed when 31d0b31 completely broke the `variable` style.

Fix it up, and add tests so this doesn't happen again.