[Fix #2582] AlignHash doesn't move keys left past the beginning of line
`Style/AlignHash` would break code under some circumstances when
autocorrecting. For example, in an effort to line up the separators, it could
move a key to left, past the beginning of the line and onto the previous line,
even overwriting the value on the previous line in the process.

There may still be some corner cases where `Style/AlignHash` breaks code.
However, this patch prevents at least one such situation.