[Fix #2549] Add ConsistentQuotesInMultiline config param for StringLiterals
Although the RuboCop developers prefer single quotes, even if it means mixing
single and double quotes in a multiline string using \ for concatenation,
many Ruby developers don't like mixed styles in a multiline string using \.

The new `ConsistentQuotesInMultiline` parameter, if set to `true`, forbids
the use of mixed single and double quotes in a multiline, continued string.
While keeping quote style consistent, it also requires that the preferred
type of quotes be used when there is no reason to use the non-preferred type.