[Fix #2448] Remove inconsistent trailing comma behavior with single item
In GitHub issue #2369, an issue was raised regarding this code:

    some_method(
      SomeObject
        .method
        .method
    )

When `Style/TrailingComma` is configured to `comma` style, it would complain
about this, because there is no trailing comma. The poster felt that this
looks strange.

The RC developers agreed that it looks strange... probably because most of us
dislike the `comma` style. But really, there is no reason why this code should
NOT have a comma, IF the cop is configured to `comma` or `consistent_comma`
styles.

If we want to make an exception for method calls/arrays/hashes with a single
item, that should have been added as a new configurable style, rather than
breaking the documented behavior of the existing `comma` style.

Therefore, restore the behavior whereby a single argument/array item/hash item
should be followed by a comma, if the relevant cop is so configured.

We do have the advantage now that `Style/TrailingComma` has been split into
`Style/TrailingCommaInArguments` and `Style/TrailingCommaInLiteral`. So if
people like trailing commas in hashes and arrays, but not in method calls,
they can easily configure RC accordingly.

Fixes #2508.