Return exit code 2 if RuboCop fails due to internal error
Previously, RC would always return 1 for failure. However, when writing scripts
which invoke RC, it is very useful to be able to distinguish "failure because
something went wrong inside RuboCop" and "failure because RuboCop found one
or more offenses".

Therefore, code 1 will mean "RuboCop found an offense", and code 2 will mean
"RuboCop terminated abnormally due to an uncaught exception".

We could further distinguish "failure because of bad CLI options",
"failure because of a bad config file", etc. but this is less useful.