Be compatible with the `RUBY_VERSION` format
This change makes it so that a user of Rubocop, can pass in
the constant `RUBY_VERSION` as the given ruby version for
`ProcessedSource`. The constant follows the format of i.e `"2.3.0"`.

This also is backwards compatible with the current behavior of passing
in a version in the format of `1.9` Or passing the version as an
`Integer` or `String`.