Performance/CaseWhenSplat reorders 'when' clauses one at a time
The next release of `parser` will include an updated `Source::Rewriter` with
adjusted rules as regards what updates "clobber" each other, and what updates
can be quietly merged. One of the adjustments is that 2 insertions cannot be
made at the exact same point. Why? Because it is impossible to know for sure
how to merge them correctly. If the same text is inserted, it is also difficult
to tell whether 1 copy or 2 should be added.

The one cop which violates this rule is Performance/CaseWhenSplat. This commit
adjusts CaseWhenSplat so that each instance will only attempt to autocorrect
a single out-of-order `when` clause.

Since autocorrection runs in a loop until there are no more corrections,
all instances of this problem will still be corrected; it may just take a
couple more iterations of the autocorrection loop.