[Fix #2775] Style/SpaceAroundKeyword accepts `yield.method`
The other keywords which can be used in this way are `super` without args (a
`zsuper` node), and `defined?`. But SpaceAroundKeyword doesn't check either of
those. A later commit will add support for them.

This adjustment will also apply to `while`, `until`, etc. But if a period
appears directly after any of those, it should result in a parse error anyways.
So I don't think it will have any effect in those cases.