Style/AndOr doesn't use AutocorrectUnlessChangingAST
There is a good reason for this; using AutocorrectUnlessChangingAST imposes
an extremely high performance cost on autocorrection. Surprisingly, after
removing it from AndOr, all the tests still pass.

Likely, AndOr was not very smart before about which expressions to flag, so
AutocorrectUnlessChangingAST was used to stop it from breaking code. Now, it may
be smarter than before.

It is also possible that there are some cases where it will perform invalid
autocorrections. If that is the case, we can fix them when they are found
(and add corresponding test cases).