Add test to check encoding when using the cache
This tests makes sure that loading from the cache returns utf8
encoded strings. Currently the encoding is being changed to
ASCII_8BIT and this can result in a fatal error for the end user.

This also modifies :location to match the pre cache-load
behaviour in lib/rubocop/processed_source.rb line 88.