Fix encoding crash on HTML output
If certain characters were encountered, RuboCop would crash during
the syntax highlighting stage. This would only happen when the
offense was read from cache.

When offenses are read from cache they come back with an ASCII-8BIT
encoding. Certain characters, such as curly quotes, cause an error
when there is an attempt to convert the string back to UTF-8 for
output to a HTML file.

This fix changes the cache serialization so that it generates
offenses in the same way as processed_source.rb, where the source
is assigned instead of being read directly from a file by Parser.