Read command line options from ./.rubocop file and RUBOCOP_OPTS env variable.
This may be useful when running rubocop using another tool, for example [pronto](https://github.com/mmozuras/pronto).

Default command-line options are loaded from `.rubocop` and `RUBOCOP_OPTS` and are combined with command-line options that are explicitly passed to `rubocop`.
Thus, the options have the following order of precedence (from highest to lowest):

1. Explicit command-line options
2. Options from `RUBOCOP_OPTS` environment variable
3. Options from `.rubocop` file.