Feature fix encoding fatal error (#3000)
* Add test for html generated from a cached result

This test also demonstrates a bug "incompatible character encodings:
UTF-8 and ASCII-8BIT". This bug only occurs when one of the scanned
files contains a certain characters and the offense results are read
from cache.

* Add test to check encoding when using the cache

This tests makes sure that loading from the cache returns utf8
encoded strings. Currently the encoding is being changed to
ASCII_8BIT and this can result in a fatal error for the end user.

This also modifies :location to match the pre cache-load
behaviour in lib/rubocop/processed_source.rb line 88.

* Fix encoding crash on HTML output

If certain characters were encountered, RuboCop would crash during
the syntax highlighting stage. This would only happen when the
offense was read from cache.

When offenses are read from cache they come back with an ASCII-8BIT
encoding. Certain characters, such as curly quotes, cause an error
when there is an attempt to convert the string back to UTF-8 for
output to a HTML file.

This fix changes the cache serialization so that it generates
offenses in the same way as processed_source.rb, where the source
is assigned instead of being read directly from a file by Parser.

* Add entry to change log for encoding fix