Fix HEREDOCs in Style/Multiline*BraceLayout Cops
bbatsov/rubocop#2873 introduced some changes to HEREDOC handling. While
these changes fixed some issues with auto-correcting and comments, they
also broke the handling of some HEREDOC use-cases.

For example, this code cannot be corrected in the symmetrical or same
line style because `EOM)` on the last line would be invalid code.

    foo(a,
      b: <<-EOM
        bar
      EOM
    )

    # invalid
    foo(a,
      b: <<-EOM
        bar
      EOM)

Additionally, this code cannot be corrected in the symmetrical or same
line style because `<<-EOM).to(<<-EOM2` would not be valid code.

    foo(<<-EOM
        bar
      EOM
    ).to(<<-EOM2
      baz
    EOM2
    )

    # invalid
    foo(<<-EOM).to(<<-EOM2
        bar
      EOM
      baz
    EOM2
    )

This change simply ignores HEREDOCs that cannot be corrected. The
determination is made by checking the HEREDOC's last line.

For example, the following can be corrected because the HEREDOC ends on
a non-conflicting line.

    foo(a,
      b: {
        c: <<-EOM
          bar
        EOM
      }
    )

    # valid
    foo(a,
      b: {
        c: <<-EOM
          bar
        EOM
      })