Handle paths where multiple potential roots exist
Update `Style/FileName` to be able to cope with files which sit at
a path in which there are multiple namespace roots present. The
specific instance this was written to deal with is our Codeship
CI run, which places the source in `~/src/`, making the full path
`~/src/project/lib/file.rb`. This resulted in failures with the
message "file.rb should define a class or module called
Project::Lib::File" because it was rooting the namespace at the
`src`.