Fix UnusedBlockArgument's IgnoreEmptyBlocks parameter from being ignored
In 8f01fbe a very old patch was applied to a new tree.  Because of this, the
changes made in that commit to the default config actually added the same key
for the `Lint/UnusedBlockArgument` cop a second time.  Because of how YAML
parses, the previous definition, which defined the `IgnoreEmptyBlocks`
parameter was lost and because of _that_, any configuration that defined it
would now throw a warning.

Also brought up in #3095