[Fix #3271] Fix bad auto-correct for `Style/EachForSimpleLoop` cop
There is a bug in `Style/EachForSimpleLoop`, where it auto-corrects
with the wrong number of iterations. Consider:

```
(0..10).each { ... }
```

This is an inclusive range of 0 through 10, i.e. 11 iterations.
This incorrectly gets auto-corrected to:

```
10.times { ... }
```

It turns out the tests for this auto-correct are incorrect, and
missing test cases for the exclusive range cases.

This fix adds test coverage, and makes the auto-correct behave
well for inclusive ranges.