Add new `Style/TernaryParentheses` cop
This cop checks the inclusion or omission of parentheses around
ternary conditions, based on its `EnforcedStyle` configuration.

Example:

```

foo = (bar && baz) ? a : b
```

will be corrected to:

```
foo = bar && baz ? a : b
```

Corrections that would change the meaning of the expression will
be avoided.

Example:

```

foo = (bar = baz = find_baz) ? a : b
```

will not be corrected.