[Fix #2645] Make `Style/EmptyLiteral` aware of frozen string literal pragma
Since in MRI >= 2.3, `String.new` and `""` are not interchangeable when
frozen string literal pragma is enabled. (The former is a mutable string,
and the latter a frozen one.)

This change makes `Style/EmptyLiteral` aware of this and doesn't generate
an offense for `String.new` when frozen string literal pragma is enabled.