[Fix #3512] Change error message of `Lint/UnneededSplatExpansion` for array (#3526)
In case of array splat for method parameter, the default message was
ambiguous. `Unnecessary splat expansion.` made people just remove the
splat expansion and pass the array as parameter which is obviously not
the same as passing splat arguments. Change the message to
`Pass array contents as separate arguments.` to
the message in such cases to make the intension more clear.