Fix disabling of autocorrect for unsafe rules by default.
This fixes PR#3398 which was supposed to disable these cops but accidentally used the wrong name for the config option.

The tests have been altered so the `AutoCorrect` config is enabled for all of them.

Also slightly changed the comment for each because _this will change to a new a method call_ doesn't make sense to me.