Fix coverage reporting
As noticed here:
https://github.com/bbatsov/rubocop/commit/f808d8da7efdeba57f11fd8b55e5fceca1071d79#commitcomment-19713872

The updated Code Climate test reporter changed the coverage percentage
from 99.2% to 0.5%. Sorry about that.

What's happening:

We're loading the SimpleCov coverage data from coverage/.resultset.json
and submitting it to Code Climate.

This repo has a really interesting test suite, which we didn't consider
in our QA. It has some tests which run in a separate process than the
main one. As configured, the test suite is telling simplecov that those
tests are part of a different test suite with a different name. Those
tests only cover a very small fraction of the code (0.5%). The main
process covers the vast majority of the code (99.2%).

A current limitation of the Code Climate ruby test reporter is that we
only submit coverage for one test suite, and don't merge the results
together. We may change that in a future version.

In the meantime, this commit acknowledges that those
in-their-own-process tests are being ignored by Code Climate, and makes
sure that the main process test suite is the one whose coverage gets
sent.