Add check on CI that manual files are in sync
I have changed the YARD rake task to be a mandatory prerequisite. The
task depends on `YARD::Registry.load!` which loads YARD information
from the .yardoc/ directory. This conditional running of YARD is part
of why the manual documentation seems to be frequently out of sync. Even
if you remember to run the rake task, your YARD documentation cache might
still be out of sync with your recent changes. Therefore, all manual
generation should be preceded by an update of they YARD cache.

The generate_cops_documentation task is skipped under jruby because the
YARD rake task fails and this results in `generate_cops_documentation`
producing empty manuals which fail the build