[Fix #3406] Enable cops if Enabled is not explicitly set
Previously, cops that have no configuration were treated as enabled by
default but cops with configuration needed to have Enabled explicitly
set. This means that adding Exclude configuration for a custom cop has
the unintuitive result of disabling it unless you also explicitly
configure it to be enabled.

This fixes #3406 by defaulting the value of Enabled to true.