Display line and column numbers when RuboCop crashes
Currently, when RuboCop crashes, it prints filename that is a target of analysis.
e.g. `An error occurred while Style/IdenticalConditionalBranches cop was inspecting /tmp/test.rb.`

I think it is not enough for debugging.
Because, we need to narrow down the cause from the whole file.
It is a laborious task.

This change makes to display line and column numbers.
e.g. `An error occurred while Style/IdenticalConditionalBranches cop was inspecting /tmp/test.rb:42:11.`

This change will make debugging easier.