[Fix #2740] Bandage on the wound of WordArray autoconfig ridiculousness
WordArray has an ill-advised feature whereby one can set a 'MinSize' config
parameter, meaning 'only enforce this rule on arrays bigger than N'. As is
common in poorly-thought-out code, thrown together without rhyme or reason,
it was not implemented correctly; the 'MinSize' rule only applies when the
enforced style is 'percent', not when it is 'brackets'. This was never
noticed, probably because the number of people using this most
questionable of features is approximately zero.

But this is just a footnote to our story of woe. No, the problem is not
that 'MinSize' does not apply to 'brackets' style. Take good heed, gentle
reader of commit logs: A feature which almost nobody wants or needs can
still break your program, even for users who don't intentionally invoke
it. Those forgotten and abandoned features are veritable breeding grounds
for bugs, legions of bugs, bugs which spill out and infect otherwise
healthy code.

In this case, when WordArray was amended to support auto-config generation,
it was made to generate a 'MinSize' config when this would help to suppress
style warnings. Why was this done? Why bother with 'MinSize'? Why not just
set the auto-generated config to {'Enabled' => false}, and be done with it?
Really, why? Mark Twain answers: 'To a man with a hammer, everything looks
like a nail.' The auto-config implementer used 'MinSize' because it was
there, not because it really helped anything.

The auto-config code had to track the size of the largest bracketed array
seen and that of the smallest percent-word array seen, to generate the
'optimal' value of 'MinSize', if anything about this sordid hatchet job
can be called 'optimal'. But it kept these values in instance variables,
and RuboCop uses a different Cop instance when checking each file. While
all the instances of WordArray shared the same 'config_to_allow_offenses'
hash, they did not share their largest/smallest array size values.

So situations arose where one WordArray instance would inspect a file
and determine that 'brackets' style should be used. Another would come
along and see a percent-word array, meaning that 'brackets' style won't
do. Since the 'EnforcedStyle' was already set to 'brackets', it would
expect that a valid array size should be stored in '@largest_brackets',
but that was not the case, because those values weren't shared between
instances.

The ugly and bad fix for this bug is to make all instances of WordArray
share their 'largest_brackets' value. Then everything "works"... but
we have more insidious global state hiding below the surface, making it
ever less likely that RuboCop can ever be properly refactored into an
independent backend library and frontend interface. (The good and
virtuous fix is to rip out the 'MinSize' parameter and consign it to
the junk heap.)