Tweak `Style/IndentHeredoc` cop
- Improve the offense message
- Update the documentation
- Fix typo in the documentation
- Fix typo in a method name

Improved messages
-------

When use Ruby 2.3 or newer

```
test.rb:2:1: C: Use 2 spaces for indentation in a heredoc by using <<~ instead of <<-.
```

When use Ruby 2.2 or older, and EnforcedStyle is not configured

```
test.rb:2:1: C: Use 2 spaces for indentation in a heredoc by using String#strip_heredoc that is provided by ActiveSupport. You need to configure EnforcedStyle to use auto-correction.
```

When use Ruby 2.2 or older, and EnforcedStyle is configured

```
test.rb:2:1: C: Use 2 spaces for indentation in a heredoc by using String#strip_heredoc that is provided by ActiveSupport.
```