`all` key in `Style/PercentLiteralDelimiters`
Oftentimes people want to set the default percent literal delimiter to [], but have to override every available option in order to get that behavior. This commit adds the ability to specify an 'all' key that will specify the default for every possible delimiter, which you can then overwrite with others.