[Fix #4001] Fix bug in Metrics/LineLength
Metrics/LineLength had a bug when configured with “ignoreCopDirectives” set to true AND checking a line with both directive # and non-directive # present.

For example:
```rb
aaa … aaa # bbb … bbb # rubocop:disable CopName
```
where (aaa … aaa # bbb … bbb) exceeds the specified line length.
The cop would correctly detect it as a violation, but would return the extraneous characters AND the directive in the error message.

```rb
LARGE_DATA_STRING_PATTERN = %r{\A([A-Za-z0-9\+\/#]…)} # rubocop:disable CopName
```
The above line has # as part of regexp and the line is longer than the permissible line length before the directive.
The cop did not detect it as a violation.