[Fix #3629] Add TargetRailsVersion functionality
As described in https://github.com/bbatsov/rubocop/issues/3629 , this
adds the option to add a TargetRailsVersion to the rubocop.yml under
AllCops. Thus, Rails-5-specific cops do not have to be manually
disengaged.

The main problem this solves is when adding RuboCop to an older Rails 4
project, running the autocorrect will actually break the specs without
much indication why.

There is a fair bit more duplication (specifically of specs) than I'd
usually like, but I figured it is better to refactor once this PR gets a
+1 from maintainers.