Fix a false negative in UselessAssignment with nested conditionals
The cop can detects the following useless assignment:

```
if flag_a
  # This is useless ...
  foo = 1
else
  # ... since this will never run together with the assignment
  puts foo
end
```

This is done by checking if branches of the assignment and reference are
exclusive.

However, previously it couldn't detect the following useless assignment:

```
if flag_a
  # This is useless ...
  foo = 1
else
  if flag_b
    # ... since this will never run together with the assignment
    puts foo
  end
end
```

Because the old logic worked correctly _only_ when the both branches
belong to a same conditional (i.e. paired `if` and `else`).

With the new logic it can detect such useless assignments by checking
ancestor branches.