Add `MinSize` option to Style/SymbolArray
In line with `Style/WordArray`, allow a `MinSize` setting to determine
the number of items to allow per array before the cop kicks in.

The minimum size handling is extracted into a new mixin, and some
other similar functionality between the two cops has been better
reflected with some method renaming and reordering.