Check for illegal combinations -P -F or -P --auto-gen-config
Running in parallel makes no sense together with --fail-fast
because only one of the processes will exit early. The other
ones run their parts of the inspection until finished.

The --auto-gen-config option requires caching to be turned off,
while --parallel requires it to be turned on. So they don't
work together.