add a message about the Style/DeprecatedHashMethods renaming
There was previously no message about the renaming of
Style/DeprecatedHashMethods to Style/PreferredHashMethods, just a
message that the cop didn't exist anymore, so I had to google to find
out the new name.

I added an entry to the OBSOLETE_COPS hash so that a message telling you
the new name is shown.