[Fix #4336] Store rubocop_cache in safer directories
A malicious person could pre-create directories and set
write access on them so that their contents could be
tampered with when RuboCop has written cache files there.

We follow the recommendations of
https://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html
to avoid these problems.

A functional difference is now is that the support for
translating `CacheRootDirectory: /tmp` to the system temporary
directory is removed. The default CacheRootDirectory is nil
(~ in YAML), which normally means $HOME/.cache, so no reason to
encourage the use of /tmp.