[Fix #4371] Prevent `Style/MethodName` from complaining about unary operator definitions
This cop would complain of not using the configured method naming scheme
when defining unary operators, e.g.:

```
def ~@
  # ...
end
```

This was happening because we internally check the method name against
our list of known operators, but the cop was passing the trailing `@` as
well, causing in a lookup miss.

This change fixes that.