[Fix #4402] Add Cop.autocorrect_incompatible_with
With this change, cops can now specify which cops their autocorrect
functionality is incompatible with. This makes it easier for third
party cops to specify cops they are not compatible with and have that
incompatibility be respected during correction. Autocorrect mapping
is not resolved on boot so this should play well with third party
extensions like rubocop-rspec.