[Fix #4419] Handle combination DisabledByDefault and Rails:Enabled
Unlike other departments, the Rails department is disabled by
default. This makes the meaning of Rails:Enabled:true different.
When AllCops:DisabledByDefault is true, setting Rails:Enabled:true
only means to remove the disablement of the department. Otherwise
it's impossible to enable the Rails cops individually (or more
precisely, it would require running with --rails).