Refactor `Lint/Syntax` cop
Why
=====

1. Apply MessageAnnotator to the cop.
2. Add the cop to the documentation.

First, currently MessageAnnotator is not enabled in `Lint/Syntax` cop.

For example:

```bash
$ rubocop --display-cop-names
Inspecting 1 file
E

Offenses:

test.rb:2:1: E: unexpected token $end
(Using Ruby 2.1 parser; configure using TargetRubyVersion parameter, under AllCops)

1 file inspected, 1 offense detected
```

`--display-cop-names` option is specified in the above example, but cop name is not displayed.

Second, `Lint/Syntax` cop does not exist in the documentation. Because the cop does not exist in `RuboCop::Cop::Cop.all`.

Purpose of the pull-request is solving the above problems.

What
=======

- Inherit `RuboCop::Cop::Cop` to `Lint/Syntax`.

We'll be able to use `add_offense` method in `Lint/Syntax` cop by the inheritance. In the method, MessageAnnotator is used. So the first problem is solved.

And by the inheritance, `Cop.all` will include `Lint/Syntax` cop. So the second problem is solved also.