Remove bang method from comparison method list
The bang method `#!` was added together with the comparison methods in
`Node::COMPARISON_OPERATORS`. This caused `SendNode#comparison_method?`
to be truthy for negation methods, which in turn led to the need for
hard-coded special cases in cops. (`YodaCondition` and
`RedundantCondition`).

This change removes `#!` from the constant, since it is in fact not a
comparison method, adds a test case to `SendNode#comparison_method?` and
simplifies `YodaCondition` based on this.