Remove duplicate node matcher macro definition from `Node`
Previously, `Node` re-implemented a near-identical version of
`NodePattern::Macros#def_node_matcher`, and in doing so directly
referenced the compiler, which is intended to be private.

This change allows `#def_node_matcher` to be called from within a node
and have the node passed as the default argument.

It also definitively makes the compiler private through
`#private_constant`.