[Fix #4101] Skip correction in Multiline*BraceLayout sometimes
If the enforced style dictates no line break before the closing
brace, but there's a comment after the last element in a list
and then a line break before the closing brace with a chained
call on it, we report the offense but don't auto-correct.

Moving the comment after the chained call could change its meaning,
so the user must do the correction manually.