Don't consider <=> for Style/YodaCondition :space_invader:
I believe, the `<=>` operator should not be considered for the yoda
condition. Maybe `.comparison_method?` shouldn't consider it.

The main reason is, that this method in difference to all other
comparison methods doesn't return a boolean, but `-1`, `0` or `1`.
Thus `0 <=> a` isn't an alternative for `a <=> 0`, because the one
returns `-1` while the other returns `1` (unless `a` is `0` of course).

I haven't checked, but I'm pretty sure the autocorrect will just
exchange the left & right side, which is wrong!