Don't consider <=> a comparison method :rocket: [Fix #4603]
I believe, the `<=>` operator should not be considered as a
`.comparison_method?`.

The main reason is, that this method in difference to all other
comparison methods doesn't return a boolean, but `-1`, `0` or `1`.
Thus, many assumptions for comparison methods don't apply, for example,
there isn't an inverse comparison method (like `<` is the inverse of
`>=)`, `* -1` isn't a comparison).

This is an alternative to #4603 by me.