Allow `has_` as a prefix for boolean methods
The [original PR][1] which suggested predicate naming for the style guide
specifically pointed out well it works with RSpec matchers:

  expect(foo).to be_red # => `foo.red?` rather than `foo.is_red?`

RSpec also has another [magic matcher][2] which does the same for have/has:

  expect(foo).to have_beans # => `foo.has_beans?`

However, the default config for Rubocop rejects this method. It would be allowed
in the spec directory, but that isn't a useful exception, because the predicate
method being tested is a domain method, not a spec method.

In addition to this removal, I would be tempted to add the other predicate
methods listed[3] in the style guide: `can` and `does`, but I think the less
controversial change is this simple one that enables behaviour that it was
always designed to do.

[1]: https://github.com/bbatsov/ruby-style-guide/pull/566
[2]: https://relishapp.com/rspec/rspec-expectations/v/2-4/docs/built-in-matchers/predicate-matchers
[3]: https://github.com/bbatsov/ruby-style-guide#bool-methods-prefix