Make `rake new_cop` create parent directories
Prior to this commit if you attempted to make a cop in a new category,
the generator would error out because the directory didn't exist yet.

This commit adds a check for the existence of the parent directory
in the `write_unless_file_exists` method and will create the directory
using `FileUtils.mkdir_p` if it doesn't exist.