Use cached remote config on network failure
If your configuration provides a URI to `inherit_from`, the locally
cached file is stale, and you are offline, then running RuboCop would
crash with an error like:

    Failed to open TCP connection to example.com:443 (getaddrinfo:
    nodename nor servname provided, or not known)

Since we do have a cached config, we can use it, and you can still run
cops without having to wait for the network to be available again.

Effectively, this behaves the same as if the network were accessible,
but the configuration file has not been modified.

If the file has never once been cached, the result is a different crash
reporting that the cached file does not exist. This may be confusing and
could be handled better, but is no worse than RuboCop's behavior when
you specify a nonexistent configuration file.