[Fix #2679] Solve LineLength:Max --auto-gen-config problem
The problem was that when --auto-gen-config was run, configuration
in .rubocop_todo.yml was not included in the run. This meant that
if there were lines longer than the default max, the setting of
Metrics/LineLength:Max in .rubocop_todo.yml caused by these lines
would not be included when running all other cops for configuration
generation.

By running in two phases we can get the correct LineLength:Max
value for cops that depend on it. The first phase puts only
Metrics/LineLength:Max in .rubocop_todo.yml. The second phase
reads that value and runs all cops using that configuration.
Finally, the Metrics/LineLength:Max from the first run is
appended to the .rubocop_todo.yml generated by the second run.

Also fixes #3496.