[Fix #4881] Fix a false positive for `HashEachMethods` when unused argument(s) exists in other blocks
Problem
===

`Performance/HashEachMethods` has a false positive for the following code.

```ruby
@h.each do |key, value|
  p(key * -value)
end
@h.sum { |key, value| -value }
```

The `each` does not have any unused arguments. `key` and `value` are used. So, this cop should not add an offense for the code.
But the cop adds an offence for the code currently.
This cause is the `sum` has an unused argument, and the argument is named `key`, the name is same as the `each`'s argument.
In this case, `Performance/HashEachMethods` adds a falsy offense.

Cause
=====

The cop handles only argument names to detect unused. So the cop does not work properly if an unused argument that has same name exists in other block.

Solution
=====

Check declared location of arguments instead of argument name.

Note
====

I found another bug while I was fixing this bug.
Auto-correction of this cop does not work properly.
For example:

```ruby
 # Auto-correct the following code.
foo.each { |key, _v| p key }
 # Expected
foo.each_key { |key| p key }
 # Actual
foo.each_value { |key| p key }
```

This change will fix the bug also.