Tweak `Lint/UnneededDisable` cop behavior.
I was noticing several weird behaviours with UnneededDisable.

First one is linked to https://github.com/bbatsov/rubocop/issues/4459,
it turns out the following piece of code:
```
  # rubocop:disable Lint/UnneededDisable
  # rubocop:disable all
```
returns the offense "Unnecessary disabling of all cops." where I believe
it shouldn't. It seems this was introduced by @rrosenblum changes
in `UnneededDisable#check` in https://github.com/bbatsov/rubocop/commit/d042ccff

Second point I noticed and wanted to address is that the following code:
```
  # rubocop:disable Lint/UnneededDisable
  # rubocop:disable Bundler/OrderedGems
  # rubocop:disable Bundler/OrderedGems
```
produces an offense for unnecessary disabling on line 3. This shouldn't be
reported as the cop is disabled. This commit addresses it as well.