[Fix #3396] Concise error when config. file not found
Instead of a stack trace, prints a friendly (concise) message
to stderr when a configuration file, e.g. inherit_from, cannot
be found.

We use the constant Errno::ENOENT because it is present on
all major operating systems. According to the ruby documenation,
there may be some odd platforms that do not support this standard.
If we find that someone is using rubocop on such a platform, we
can rescue SystemCallError instead, but that will make it hard
to exit with the correct exit code.