[Refactor] Move resolution related code to ConfigLoaderResolver
Move related functionality into one place. Try a different
approach than throwing stuff out into a mixin.

Make ConfigLoaderResolver a class because that is cleaner than
mixing in the module. A mixin and the class that mixes in the
module have access to each other's private members. Also,
method calls without a receiver go to either a method in the
class or in the module, which can be confusing.

Now there's some calling back and forth between ConfigLoader and
ConfigLoaderResolver, which isn't a perfect solution, but at least
we get some explicit receivers, which should improve readability.