Restyle the example formatting for Naming cops
Follow up of https://github.com/bbatsov/rubocop/pull/4880#issuecomment-338499947.

This commit is a change of document format. Cops of `naming` department
are the target.

Since example of `Naming/HeredocDelimiterCase` cop was the opposite of
`bad` and `good`, it was rearranged for unification.
Also added `default` information based on `config/default.yml`.