[Fix #4892] Add new `Lint/ShadowedArgument` cop (#4908)
Since recently, unused argument cops (`Lint/UnusedBlockArgument` and
`Lint/UnusedMethodArgument`) handle the case of shadowed
arguments. However, shadowed arguments detection functionality differs
semantically from unused arguments cops, and, as such, should be
extracted to a separate cop.

This change
* Introduces `Lint/ShadowedArgument` cop.
* Makes `Cop::VariableForce::Assignment` capable of tracking
  assignment's references.
* Removes the shadowed argument detection algorithm from unused
  argument cops.

## `Lint/ShadowedArgument` cop.

### The algorithm
The shadowed argument detection algorithm is slightly improved, so it
handles the case when reassignment occurs inside blocks or conditionals
of arbitrary nesting. The algoritm takes in account argument
reassigning in blocks (fixes #4898).

The algorithm searches for reassignment nodes and reports offenses on them
if an argument was reassigned before it was referenced.
In case if the argument  was reassigned inside a block or conditional, the
algorithm signals that precise offense location is undecidable (since
it's unknown whether an entity is executed). In this case, offense
is reported on argument's declaration node.

### Configuration
It's reported (#4892) that in real-life code there may be situtations
when argument shadowing is used in order to pass
parameters to `zsuper`:

```ruby
def foo(bar)
  bar = 42
  super
end
```

If authors tend to use this style, they can use the
`IgnoreImplicitReferences` configuration option.