[Fix 4363] Fix false offenses detected by `Style/PercentLiteralDelimiters` cop
Fixes #4363.

Percent notation targeted by Issue # 4363 are `%w` and `%i`.
`%W` and `%I` are not targets.

## %w / %W

The meaning changes.

```ruby
> %w(\() # => ["("]
> %w[\(] # => ["\\("]
```

The meaning does not change.

```ruby
> %W(\() # => ["("]
> %W[\(] # => ["("]
```

## %i / %I

The meaning changes.

```ruby
> %i(\() # => [:"("]
> %i[\(] # => [:"\\("]
```

The meaning does not change

```ruby
> %I(\() # => [:"("]
> %I[\(] # => [:"("]
```

Also, if we use open delimiter and closing delimiter like `(...)`, we
need to target both of them.

```ruby
> %w(\() # => ["("]
> %w[\(] # => ["\\("]
> %w(\)) # => [")"]
> %w[\)] # => ["\\)"]
```

There are 4 cases of `%w(...)`, `%w[...]`, `%w{...}`, and `%w<...>`.

This commit changes so that it will not be offense if the same character
as a percent notation delimiter is used for elements of an array.