Drop rbx-3 from travis.yml
## Summary

rbx-3 has been failed in Travis CI for a long while.
It seems that it already failed when upgrading from rbx-2 to rbx-3.
https://github.com/bbatsov/rubocop/pull/2706

It seems that there is no activity while failing with `allow_failures`.
So I think that it is not necessary to use CI resources at present.
When passing the test with rbx-3, I think that it is better to add it
again to travis.yml.

## Other Information

jruby-head (jruby 9.2.0.0-SNAPSHOT) will progress as soon as the
following Issue is resolved.
https://github.com/jruby/jruby/issues/4809