[Fix #4662] Split heredoc indentation by lines
Checking indentation of heredocs would fail if the first line was empty,
with a nonsense message:

    <<~HEREDOC

      foo
    HEREDOC

    # Use 2 spaces for indentation in a heredoc by using `<<~` instead of `<<~`.

Indentations were detected as leading contiguous whitespace. In the
example above, this would be "\n  ", which fails the check due to being
3 characters long.

By first splitting lines, we are can safely rely on the leading
contiguous whitespace being the indent.

Resolves: #4662
See also: #4465