[Fix #4449] Make `Layout/IndentHeredoc` aware of line length
Problem
===

See #4449

`Layout/IndentHeredoc` does not treat line length. So the cop adds an offense for a heredoc with long line.

For example

```ruby
module A
  module B
    class C
      def foo
        puts <<-MSG
This message's max line length is 75.
So, if you add indentations to this heredoc, this code violates LineLength.
But if `Metrics/LineLength`'s `AllowHeredoc` is true(default),
it does not violate.
        MSG
      end
    end
  end
end
```

Solution
=====

This change will make `Layout/IndentHeredoc` to not register any offenses for a heredoc with long line when `AllowHeredoc` is false.