Make new_cop task to modify `config/enabled.yml`
# Problem

Currently, new_cop task does not modify `config/enabled.yml`(or `config/disabled.yml`).
It is not good. We should edit `config/enabled.yml` manually when we add a new cop.

# Solution

Make new_cop task to modify `config/enabled.yml`.
By this change, we should edit only a description for an added cop in `config/enabled.yml`.
When we want to disable a cop by default, we need to move the added configuration to `config/disabled.yml`. But added cop is enabled by default in almost cases, so I think this will not be a problem.